Move to "Internal" namespace
