+++ Source/Testably.Abstractions.Testing/FileSystemInitializer.Initializer.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Testably.Abstractions.Testing.Internal;
 
 namespace Testably.Abstractions.Testing;
