+++ Source/Testably.Abstractions.Testing/FileSystemMock.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Testably.Abstractions.Testing.Internal;
 using Testably.Abstractions.Testing.Storage;
 
 namespace Testably.Abstractions.Testing;
