+++ Tests/Testably.Abstractions.Testing.Tests/FileSystemMockTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using Testably.Abstractions.Testing.Internal;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
 namespace Testably.Abstractions.Testing.Tests;
