+++ Tests/Testably.Abstractions.Testing.Tests/Internal/RandomSystemExtensionsTests.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Testably.Abstractions.Testing.Internal;
 
-namespace Testably.Abstractions.Testing.Tests;
+namespace Testably.Abstractions.Testing.Tests.Internal;
 
 public class RandomSystemExtensionsTests
 {
